suw: Disable ScreenLockActivity if a secure lockscreen is already set
Currently, devices that have a fingerprint prompt to add a fingerprint and
lockscreen via FingerprintActivity. However, since ScreenLockActivity is
always active, we end up with a second prompt for screen lock security.

This adds a check to see if a secure lockscreen has already been set, and
if so, just skip ScreenLockActivity.

Change-Id: I877835e8fa8628c01c96ab9732146720c1234f31